+++ apps/webview/src/components/Layout/TabBar.tsx
@@ -1,9 +1,10 @@
-import { css } from '_panda/css';
+import { css, cx } from '_panda/css';
 import { flex } from '_panda/patterns';
 import { wrap } from '@suspensive/react';
 import { GamepadIcon, HouseIcon, StoreIcon, UserRoundIcon } from 'lucide-react';
 
 import { Link, usePathname } from '@/i18n/routing';
+import { containerStyle } from '@/styles/layout';
 
 const TAB_LIST = [
   {
@@ -39,21 +40,24 @@ export const TabBar = wrap.Suspense().on(() => {
   const pathname = usePathname();
   return (
     <div
-      className={css({
-        display: 'grid',
-        gridTemplateColumns: 'repeat(4, 1fr)',
-        gap: '18px',
-        pt: 3,
-        pb: 1,
-        px: 10,
-        position: 'fixed',
-        bottom: 0,
-        left: 0,
-        right: 0,
-        zIndex: 100,
-        backgroundColor: '#212429',
-        borderRadius: '16px 16px 0px 0px',
-      })}
+      className={cx(
+        containerStyle,
+        css({
+          display: 'grid',
+          gridTemplateColumns: 'repeat(4, 1fr)',
+          gap: '18px',
+          pt: 3,
+          pb: 1,
+          px: 10,
+          position: 'fixed',
+          bottom: 0,
+          left: 0,
+          right: 0,
+          zIndex: 100,
+          backgroundColor: '#212429',
+          borderRadius: '16px 16px 0px 0px',
+        }),
+      )}
     >
       {TAB_LIST.map((tab) => (
         <Link
